# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Smart Web Scraper API
  description: 지능형 웹 데이터 수집 플랫폼 — 5단계 적응형 파이프라인 + Self-Healing
  version: 1.0.0
  contact:
    name: Smart Web Scraper
servers:
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /api/v1/urls:
    get:
      tags: [URLs]
      summary: URL 목록 조회
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: size
          in: query
          schema: { type: integer, default: 20 }
        - name: search
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
        - name: is_active
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: URL 목록
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/UrlResponse" }
                  total: { type: integer }
                  page: { type: integer }
                  pages: { type: integer }

    post:
      tags: [URLs]
      summary: URL 등록 (단일 또는 일괄)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/UrlCreate"
                - $ref: "#/components/schemas/BulkUrlCreate"
      responses:
        "201": { description: 등록 성공 }
        "403": { description: robots.txt 차단 }
        "409": { description: 중복 URL }
        "422": { description: 유효하지 않은 URL }
  /api/v1/urls/import:
    post:
      tags: [URLs]
      summary: CSV 파일로 URL 일괄 등록
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: 가져오기 결과
          content:
            application/json:
              schema:
                type: object
                properties:
                  imported: { type: integer }
                  errors: { type: array, items: { type: string } }
  /api/v1/urls/{id}:
    get:
      tags: [URLs]
      summary: URL 상세 조회
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UrlResponse" }
        "404": { description: URL 없음 }

    put:
      tags: [URLs]
      summary: URL 수정
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UrlUpdate" }
      responses:
        "200": { description: 수정 성공 }

    delete:
      tags: [URLs]
      summary: URL 삭제
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: 삭제 성공 }

  # ─── Scrape ───────────────────────────────────────────────
...
components:
  schemas:
    UrlCreate:
      type: object
      required: [url]
      properties:
        url: { type: string, format: uri }
        name: { type: string }
        category: { type: string }
        extraction_schema: { type: object }
        tags: { type: array, items: { type: string } }
    BulkUrlCreate:
      type: object
      required: [urls]
      properties:
        urls:
          type: array
          items:
            type: object
            required: [url]
            properties:
              url: { type: string, format: uri }
              name: { type: string }
    UrlUpdate:
      type: object
      properties:
        name: { type: string }
        is_active: { type: boolean }
        extraction_schema: { type: object }
        tags: { type: array, items: { type: string } }
    UrlResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        url: { type: string }
        name: { type: string }
        domain: { type: string }
        category: { type: string }
        is_active: { type: boolean }
        tags: { type: array, items: { type: string } }
        created_at: { type: string, format: date-time }
        profile: { $ref: "#/components/schemas/UrlProfile" }
    UrlProfile:
      type: object
      properties:
        best_pipeline: { type: string }
        success_rate: { type: number }
        total_visits: { type: integer }
        avg_response_time_ms: { type: number }
        requires_js: { type: boolean }
        has_antibot: { type: boolean }
        antibot_type: { type: string }
...